Restructured plot class structure
