+++ GuiFunctions/MetaDraw/SpectrumMatch/CrosslinkSpectrumMatchPlot.cs
@@ -7,10 +7,10 @@ namespace GuiFunctions
     /// <summary>
     /// Class for displaying CrossLinked spectra in MetaDraw
     /// </summary>
-    public class CrosslinkSpectrumMatchPlot : PeptideSpectrumMatchPlot
+    public class CrosslinkSpectrumMatchPlot : SpectrumMatchPlot
     {
         public CrosslinkSpectrumMatchPlot(OxyPlot.Wpf.PlotView plotView, PsmFromTsv csm, MsDataScan scan, Canvas stationaryCanvas)
-            : base(plotView, csm, scan, csm.MatchedIons)
+            : base(plotView, csm, scan)
         {
             // annotate beta peptide matched ions
             AnnotateMatchedIons(isBetaPeptide: true, csm.BetaPeptideMatchedIons);
